move all signature aggregeation logic to AggregateSignatures
